Changed structure for Maven Support
