--- HealthBar/.classpath
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="lib" path="C:/Users/Filippo/Eclipse/Librerie/bukkit-1.6.4-R0.1-20130923.182036-4.jar"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="C:/Users/Filippo/Eclipse/Librerie/EpicBossGoldEdition.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
