--- HealthBar/src/com/gmail/filoghost/healthbar/BarType.java
@@ -1,9 +1,0 @@
-package com.gmail.filoghost.healthbar;
-
-public enum BarType {
-	
-	BAR, 			//this means that the bars are stored in an array, custom or not
-	DEFAULT_TEXT,	//this needs some string replacements
-	CUSTOM_TEXT 	//this needs a lot of string replacements
-	
-}
