+++ src/main/java/com/gmail/filoghost/healthbar/StringBoolean.java
@@ -1,0 +1,20 @@
+package com.gmail.filoghost.healthbar;
+
+public class StringBoolean {
+
+	private String s;
+	private Boolean b;
+	
+	public StringBoolean (String s, Boolean b) {
+		this.s = s;
+		this.b = b;
+	}
+	
+	public String getString() {
+		return s;
+	}
+	
+	public Boolean getBoolean() {
+		return b;
+	}
+}
