+++ src/main/resources/locale.yml
@@ -1,0 +1,28 @@
+CREEPER: Creeper
+SKELETON: Skeleton
+SPIDER: Spider
+GIANT: Giant
+ZOMBIE: Zombie
+SLIME: Slime
+GHAST: Ghast
+PIG_ZOMBIE: Zombie Pigman
+ENDERMAN: Enderman
+CAVE_SPIDER: Cave Spider
+SILVERFISH: Silverfish
+BLAZE: Blaze
+MAGMA_CUBE: Magma Cube
+ENDER_DRAGON: Ender Dragon
+WITHER: Wither
+BAT: Bat
+WITCH: Witch
+PIG: Pig
+SHEEP: Sheep
+COW: Cow
+CHICKEN: Chicken
+SQUID: Squid
+WOLF: Wolf
+MUSHROOM_COW: Mooshroom
+SNOWMAN: Snowman
+OCELOT: Ocelot
+IRON_GOLEM: Iron Golem
+VILLAGER: Villager
